Deprecate VoronoiFVM.Physics callbacks without `data` argument
This patch implements a warning that is triggered if the user constructs
a VoronoiFVM.Physics object with a callback function that uses only three or
two arguments, respectively.

All examples were overhauled to comply with the new interface.

A MaybeData dummy struct was added.

This helper struct represents data or not.
The purpose is to control whether we call the Physics callbacks with or without data.
A heuristic check is performed to detect the signature of the callbacks.
This can safely be removed once the deprecation phase is over.

With MaybeData the dummies nofunc2, nosrc2 and default_storage2 can be removed.